core/requirements.server.txt
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
attrs==25.3.0
Authlib==1.3.2
backports.tarfile==1.2.0
beautifulsoup4==4.13.5
black==25.9.0
boto3==1.39.11
botocore==1.39.11
bs4==0.0.2
build==1.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
colorlog==6.9.0
coverage==7.10.7
cryptography==46.0.1
dataclasses-json==0.6.7
datasets==4.1.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docutils==0.22.2
et_xmlfile==2.0.0
expandvars==1.1.2
fastapi==0.115.14
filelock==3.19.1
filetype==1.2.0
# -e git+ssh://git@github.com/h2oai/flood_prediction.git@2136147bf81b288fb64e618bf5b816219e565ab1#egg=flood_prediction&subdirectory=core
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
h2o-authn==3.1.0
h2ogpte==1.6.41rc3
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
langchain==0.3.27
langchain-aws==0.2.33
langchain-community==0.3.29
langchain-core==0.3.76
langchain-milvus==0.2.1
langchain-nvidia-ai-endpoints==0.3.18
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.29
lxml==6.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.14.1
mcp-server-time==2025.8.4
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.0
nodeenv==1.9.1
numpy==1.26.4
nvidia-nat==1.3.0a20250917
nvidia-nat-langchain==1.3.0a20250917
nvidia-nat-mcp==1.3.0a20250917
openai==1.107.1
openinference-semantic-conventions==0.1.21
openpyxl==3.1.5
optuna==4.4.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pkce==1.0.3
pkginfo==1.12.1.2
platformdirs==4.4.0
pluggy==1.6.0
ply==3.11
propcache==0.3.2
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymilvus==2.6.2
pyproject_hooks==1.2.0
pyright==1.1.405
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
ragas==0.2.15
readme_renderer==44.0
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.27.1
rq==2.1.0
s3transfer==0.13.1
SecretStorage==3.4.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.46.2
structlog==23.2.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
tomli==2.2.1
tqdm==4.67.1
twine==6.2.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==11.0.3
wikipedia==1.4.0
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
